fix(config): align env schema validation
